\name{calcCcafs.logical}
\alias{calcCcafs.logical}
\title{Eucledian distance to calculate dissimilarity...}
\usage{\method{calcCcafs}{logical} (lag, params, training, weights, base,
    project)}
\description{Eucledian distance to calculate dissimilarity}
\arguments{\item{params}{an object of the class AnalogueParameters}
\item{...}{more to come}}
\examples{ccafs_params <- dissimilarity(x, z, )}
