\name{calcHal.logical}
\alias{calcHal.logical}
\title{Method after Hallegatte...}
\usage{\method{calcHal}{logic} (lag, params, training, weights, base, project)}
\description{Method after Hallegatte}
\arguments{\item{params}{an object of the class AnalogueParameters}
\item{...}{more to come}}
\examples{ccafs_params <- dissimilarity(x, z, )}
