\name{loadData}
\alias{loadData}
\title{Load data for the dissimilarity calculations...}
\usage{loadData(params)}
\description{Load data for the dissimilarity calculations}
\details{This functions loads the dissimilarity data.}
\keyword{manip}
\value{an object of AnalogueTraining}
\arguments{\item{params}{an object of the class AnalogueParameters}}
\examples{ccafs_params <- createParams(x, z, )}
