\name{loadWeights}
\alias{loadWeights}
\title{Load weights for grids...}
\usage{loadWeights(params)}
\description{Load weights for grids}
\details{This functions loads weight grids, if they exist. Otherwise
weights are set to 1 or the value specified.}
\value{an object of class AnalogueWeights}
\arguments{\item{params}{an object of the class AnalogueParameters}}
