\name{makeRef.ATraining}
\alias{makeRef.ATraining}
\title{Extract reference from data grid...}
\usage{makeRef.ATraining(data, params)}
\description{Extract reference from data grid}
\value{an object of AnalogueResults}
\arguments{\item{params}{an object of the class AnalogueParameters}
\item{weights}{a list with all the weights}}
\examples{ccafs_params <- dissimilarity(x, z, )}
