\name{makeRef}
\alias{makeRef}
\title{Extract reference values from rasters...}
\usage{makeRef(object, ...)}
\description{Extract reference values from rasters}
\value{an object of AnalogueResults}
\arguments{\item{params}{an object of the class AnalogueParameters}
\item{...}{an object of the class AnalogueTraining, AnalogueTraining}}
\examples{ccafs_params <- dissimilarity(x, z, )}
